Modified write process.
Rather than writing a single JSON file, each sample is written to its own file in a newly-created directory. The directory is then compressed into a gzipped tarball. The file names are UUIDs generated from the MD5 checksum of the JSON file.

Duplicates are disallowed, and are identified when their checksums match.

By default, a duplicate will abort the conversion and clean up the remaining file. This behavior can be made explicit with the --duplicate-error flag.

To skip duplicates, invoke pifify with the --duplicate-warning flag.