Refactored pifify to separate sample definitions and handle input that allows for more ready expansion to additional builds and to additional suppliers.
This commit preserves the function `csv_kernel` in `pifify.py`, which reads CSV input output from `part_enum`, an initial, and hackish way of transforming the build data into a CSV. This method is very fragile, since it relies on an intimate knowledge between the CSV writer and the (separate) CSV reader here. The next step is to remove the CSV reader. It may be reintroduced later, but now there is no need.

The anticipated use case is for pifify to be called with keywords, e.g. faustson plate 1, or faustson plate 2.